+++ types/event.go
@@ -39,3 +39,42 @@ func (e *Event) Validate() error {
 
 	return nil
 }
+
+// HasMetrics determines if an event has metric data.
+func (e *Event) HasMetrics() bool {
+	if e.Metrics != nil {
+		return true
+	}
+
+	return false
+}
+
+// IsIncident determines if an event indicates an incident.
+func (e *Event) IsIncident() bool {
+	if e.Check.Status != 0 {
+		return true
+	}
+
+	return false
+}
+
+// IsResolution returns true if an event has just transitionned from an incident
+func (e *Event) IsResolution() bool {
+	// Try to retrieve the previous status in the check history and verify if it
+	// was a non-zero status, therefore indicating a resolution
+	if len(e.Check.History) > 0 && e.Check.History[len(e.Check.History)-1].Status != 0 &&
+		!e.IsIncident() {
+		return true
+	}
+
+	return false
+}
+
+// IsSilenced determines if an event has any silenced entries
+func (e *Event) IsSilenced() bool {
+	if len(e.Silenced) > 0 {
+		return true
+	}
+
+	return false
+}